Add transfer and init functions for the IMU

+++ include/imu.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void IMU_SPI_Init();
-uint16_t IMU_SPI_Transfer(uint32_t MOSI);
+uint16_t IMU_SPI_Transfer(uint8_t MOSI);
 uint16_t IMU_Read_Register(uint8_t RegAddr);
 uint16_t IMU_Write_Register(uint8_t RegAddr, uint8_t RegValue);
 void IMU_Update_SPI_Config();
